create components for cases in tx view
