+++ src/components/swap/swap-modals/TxView.js
@@ -1,27 +1,15 @@
-import React, { useContext, useEffect, useState } from 'react';
+import React, { useContext, useState } from 'react';
 import styled, { css } from 'styled-components/macro';
-import { Icon, Divider } from 'semantic-ui-react';
-import { LightModeContext } from '../../../contexts/LightModeContext';
 import { useGameEditionContext } from '../../../contexts';
-import { AccountContext } from '../../../contexts/AccountContext';
-import { ChainIcon, ErrorIcon, PixeledCircleArrowIcon, SuccessfullIcon } from '../../../assets';
-import { GameEditionContext, GE_DESKTOP_CONFIGURATION } from '../../../contexts/GameEditionContext';
+import { ErrorIcon } from '../../../assets';
+import { GameEditionContext } from '../../../contexts/GameEditionContext';
 import { SwapContext } from '../../../contexts/SwapContext';
-import { extractDecimal, gasUnit, reduceBalance } from '../../../utils/reduceBalance';
 import CustomButton from '../../../components/shared/CustomButton';
-import reduceToken from '../../../utils/reduceToken';
-import { PactContext } from '../../../contexts/PactContext';
-import tokenData from '../../../constants/cryptoCurrencies';
-import PopupTxView from './PopupTxView';
-import { commonColors, theme } from '../../../styles/theme';
 import Label from '../../shared/Label';
-import { ENABLE_GAS_STATION, GAS_PRICE } from '../../../constants/contextConstants';
 import { LIQUIDITY_VIEW } from '../../../constants/liquidityView';
-import GameEditionLabel from '../../game-edition-v2/components/GameEditionLabel';
-import { InfoContainer } from '../../game-edition-v2/components/PixeledInfoContainerBlue';
-import PressButtonToActionLabel from '../../game-edition-v2/components/PressButtonToActionLabel';
-import { PixeledInfoContainerWhite } from '../../game-edition-v2/components/PixeledInfoContainerWhite';
-import PixeledBlueContainer from '../../game-edition-v2/components/PixeledInfoContainerBlue';
+import { SuccessAddRemoveView, SuccessAddRemoveViewGE } from './LiquidityTxView';
+import { SwapSuccessView, SwapSuccessViewGE } from './SwapSuccesTxView';
+import { commonColors } from '../../../styles/theme';
 
 const Content = styled.div`
   display: flex;
@@ -64,255 +52,17 @@ const TransactionsDetails = styled.div`
   }
 `;
 
-const Row = styled.div`
-  display: flex;
-  align-items: center;
-  &.sb {
-    justify-content: space-between;
-  }
-  &.fs {
-    justify-content: flex-start;
-  }
-  &.c {
-    justify-content: center;
-  }
-`;
-
-const PreviewContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  & > *:not(:last-child) {
-    margin-bottom: 16px;
-  }
-`;
-
 const TxView = ({ view, onClose, token0, token1, createTokenPair }) => {
   const swap = useContext(SwapContext);
   const { gameEditionView } = useContext(GameEditionContext);
-  const { account } = useContext(AccountContext);
-  const { themeMode } = useContext(LightModeContext);
-  const pact = useContext(PactContext);
 
   const [loading, setLoading] = useState(false);
-
-  const showTicker = (ticker) => {
-    if (ticker === 'coin') return 'KDA';
-    else if (ticker === 'runonflux.flux') return 'FLUX';
-    else return ticker?.toUpperCase();
-  };
-
-  const getTokenIcon = (token) => {
-    return tokenData[showTicker(token)]?.icon;
-  };
 
   const sendTransaction = () => {
     setLoading(true);
     swap.swapSend();
     onClose();
     setLoading(false);
-  };
-
-  const successViewGE = () => {
-    return (
-      <SuccessViewContainerGE
-        leftItem={
-          <>
-            <Row className="fs">
-              {getTokenIcon(swap?.localRes?.result?.data[0]?.token)}
-              <GameEditionLabel fontSize={32} color="black" fontFamily="bold">
-                {extractDecimal(swap?.localRes?.result?.data[0]?.amount)}
-              </GameEditionLabel>
-            </Row>
-
-            <GameEditionLabel color="blue">From</GameEditionLabel>
-            <Row className="fs">
-              <GameEditionLabel fontSize={22} color="blue-grey">
-                {reduceToken(account.account)}
-              </GameEditionLabel>
-              <ChainIcon className="chain-icon" />
-              <GameEditionLabel fontSize={22} color="blue-grey">
-                {swap?.localRes?.metaData?.publicMeta?.chainId}
-              </GameEditionLabel>
-            </Row>
-          </>
-        }
-        rightItem={
-          <>
-            <Row className="fs">
-              {getTokenIcon(swap?.localRes?.result?.data[1]?.token)}
-              <GameEditionLabel fontSize={32} color="black" fontFamily="bold">
-                {extractDecimal(swap?.localRes?.result?.data[1]?.amount)}
-              </GameEditionLabel>
-            </Row>
-            <GameEditionLabel color="blue">From</GameEditionLabel>
-            <Row className="fs">
-              <GameEditionLabel fontSize={22} color="blue-grey">
-                {reduceToken(account.account)}
-              </GameEditionLabel>
-              <ChainIcon className="chain-icon" />
-              <GameEditionLabel fontSize={22} color="blue-grey">
-                {swap?.localRes?.metaData?.publicMeta?.chainId}
-              </GameEditionLabel>
-            </Row>
-          </>
-        }
-        infoItems={[
-          {
-            label: `${showTicker(swap?.localRes?.result?.data[0]?.token)}/${showTicker(swap?.localRes?.result?.data[1]?.token)}`,
-            value: `1 = ${reduceBalance(pact?.computeOut(1), 12)}`,
-          },
-          {
-            label: 'gas cost KDA',
-            value: ENABLE_GAS_STATION ? (
-              <>
-                <GameEditionLabel geColor="white">{gasUnit(GAS_PRICE * swap?.localRes?.gas)} KDA</GameEditionLabel>
-                <GameEditionLabel geColor="white" labelStyle={{ marginLeft: 5 }}>
-                  FREE!
-                </GameEditionLabel>
-              </>
-            ) : (
-              <GameEditionLabel geColor="white">{gasUnit(GAS_PRICE * swap?.localRes?.gas)} KDA</GameEditionLabel>
-            ),
-          },
-        ]}
-      />
-    );
-  };
-
-  const successView = () => {
-    return (
-      <SuccesViewContainer
-        swap={swap}
-        loading={loading}
-        onClick={() => {
-          sendTransaction();
-        }}
-      >
-        <Row className="sb">
-          <Label fontFamily="bold">From</Label>
-          <Label fontSize={13}></Label>
-        </Row>
-        <Row className="sb">
-          <Label fontSize={13}>Account</Label>
-          <Label fontSize={13}>
-            {`${reduceToken(account.account)}`}
-            <PopupTxView isAccountPopup />
-          </Label>
-        </Row>
-        <Row className="sb">
-          <Label fontSize={13}>Chain ID</Label>
-          <Label fontSize={13}>{swap?.localRes?.metaData?.publicMeta?.chainId}</Label>
-        </Row>
-        <Divider
-          style={{
-            width: '100%',
-            marginTop: 0,
-            borderTop: gameEditionView ? `2px dashed ${theme(themeMode).colors.black}` : `1px solid ${theme(themeMode).colors.white}`,
-          }}
-        />
-
-        <Row className="sb">
-          <Row className="fs">
-            <Label fontFamily="bold"> {getTokenIcon(swap?.localRes?.result?.data[0]?.token)}</Label>
-            <Label fontFamily="bold">{extractDecimal(swap?.localRes?.result?.data[0]?.amount)}</Label>
-          </Row>
-          <Label fontFamily="bold">{showTicker(swap?.localRes?.result?.data[0]?.token)}</Label>
-        </Row>
-        <Row className="sb">
-          <Label fontFamily="bold">
-            <Icon name="long arrow alternate down" style={{}} />
-          </Label>
-          <Label fontSize={13}>{`1 ${showTicker(swap?.localRes?.result?.data[0]?.token)} = ${reduceBalance(pact?.computeOut(1), 12)} ${showTicker(
-            swap?.localRes?.result?.data[1]?.token
-          )}`}</Label>
-        </Row>
-        <Row className="sb">
-          <Row className="fs">
-            <Label fontFamily="bold"> {getTokenIcon(swap?.localRes?.result?.data[1]?.token)}</Label>
-            <Label fontFamily="bold">{`${extractDecimal(swap?.localRes?.result?.data[1]?.amount)} `}</Label>
-          </Row>
-          <Label fontFamily="bold">{` ${showTicker(swap?.localRes?.result?.data[1]?.token)}`}</Label>
-        </Row>
-      </SuccesViewContainer>
-    );
-  };
-
-  const successAddRemoveViewGE = ({ label }) => {
-    return (
-      <SuccessViewContainerGE
-        hideIcon
-        title={label}
-        leftItem={
-          <>
-            <GameEditionLabel fontSize={32} color="blue">
-              {showTicker(token0)}
-            </GameEditionLabel>
-            <Row className="fs">
-              {getTokenIcon(token0)}
-              <GameEditionLabel fontSize={22} color="blue-grey">
-                {extractDecimal(swap?.localRes?.result?.data?.amount0)}
-              </GameEditionLabel>
-            </Row>
-          </>
-        }
-        rightItem={
-          <>
-            <GameEditionLabel fontSize={32} color="blue">
-              {showTicker(token1)}
-            </GameEditionLabel>
-
-            <Row className="fs">
-              {getTokenIcon(token1)}
-              <GameEditionLabel fontSize={22} color="blue-grey">
-                {extractDecimal(swap?.localRes?.result?.data?.amount1)}
-              </GameEditionLabel>
-            </Row>
-          </>
-        }
-        infoItems={[
-          {
-            label: 'gas cost KDA',
-            value: ENABLE_GAS_STATION ? (
-              <>
-                <GameEditionLabel geColor="white">{gasUnit(GAS_PRICE * swap?.localRes?.gas)} KDA</GameEditionLabel>
-                <GameEditionLabel geColor="white" labelStyle={{ marginLeft: 5 }}>
-                  FREE!
-                </GameEditionLabel>
-              </>
-            ) : (
-              <GameEditionLabel geColor="white">{gasUnit(GAS_PRICE * swap?.localRes?.gas)} KDA</GameEditionLabel>
-            ),
-          },
-        ]}
-      />
-    );
-  };
-
-  const successAddRemoveView = ({ label, onClick }) => {
-    return (
-      <SuccesViewContainer swap={swap} loading={loading} onClick={onClick}>
-        <Row className="fs">
-          <Label fontFamily="bold">{label}</Label>
-        </Row>
-        <Row className="sb">
-          <Row className="fs">
-            {getTokenIcon(token0)}
-            <Label fontFamily="bold">{extractDecimal(swap?.localRes?.result?.data?.amount0)}</Label>
-          </Row>
-          <Label fontFamily="bold">{showTicker(token0)}</Label>
-        </Row>
-        <Row className="fs">
-          <Label fontFamily="bold">{label}</Label>
-        </Row>
-        <Row className="sb">
-          <Row className="fs">
-            {getTokenIcon(token1)}
-            <Label fontFamily="bold">{extractDecimal(swap?.localRes?.result?.data?.amount1)}</Label>
-          </Row>
-          <Label fontFamily="bold">{showTicker(token1)}</Label>
-        </Row>
-      </SuccesViewContainer>
-    );
   };
 
   const failView = () => {
@@ -396,11 +146,23 @@ const TxView = ({ view, onClose, token0, token1, createTokenPair }) => {
         default:
           return () => {};
         case LIQUIDITY_VIEW.REMOVE_LIQUIDITY:
-          return gameEditionView ? successAddRemoveViewGE({ label: 'Remove' }) : successAddRemoveView({ label: 'Remove', onClick: sendTransaction });
+          return gameEditionView ? (
+            <SuccessAddRemoveViewGE token0={token0} token1={token1} swap={swap} label="Remove" />
+          ) : (
+            <SuccessAddRemoveView token0={token0} token1={token1} swap={swap} label="Remove" loading={loading} onClick={sendTransaction} />
+          );
         case LIQUIDITY_VIEW.ADD_LIQUIDITY:
-          return gameEditionView ? successAddRemoveViewGE({ label: 'Add' }) : successAddRemoveView({ label: 'Add', onClick: onAddLiquidity });
+          return gameEditionView ? (
+            <SuccessAddRemoveViewGE token0={token0} token1={token1} swap={swap} label="Add" />
+          ) : (
+            <SuccessAddRemoveView token0={token0} token1={token1} swap={swap} label="Add" loading={loading} onClick={onAddLiquidity} />
+          );
         case undefined:
-          return gameEditionView ? successViewGE() : successView();
+          return gameEditionView ? (
+            <SwapSuccessViewGE swap={swap} />
+          ) : (
+            <SwapSuccessView swap={swap} loading={loading} sendTransaction={sendTransaction} />
+          );
       }
     } else return failView();
   };
@@ -433,93 +195,5 @@ const Message = ({ color, children }) => {
         {children}
       </Label>
     </MessageContainer>
-  );
-};
-
-// GAS COST COMPONENT
-const GasCost = ({ swap }) => {
-  return (
-    <Row className="sb">
-      <Label fontSize={13}>Gas Cost</Label>
-      <div style={{ display: 'flex' }}>
-        {ENABLE_GAS_STATION ? (
-          <>
-            <Label fontSize={13} color={commonColors.green} geColor="green">
-              {gasUnit(GAS_PRICE * swap?.localRes?.gas)} KDA
-            </Label>
-            <Label fontSize={13} color={commonColors.green} geColor="green" labelStyle={{ marginLeft: 5 }}>
-              FREE!
-            </Label>
-          </>
-        ) : (
-          <Label fontSize={13} color={commonColors.green} geColor="green">
-            {gasUnit(GAS_PRICE * swap?.localRes?.gas)} KDA
-          </Label>
-        )}
-        {ENABLE_GAS_STATION && <PopupTxView popupStyle={{ maxWidth: '400px' }} />}
-      </div>
-    </Row>
-  );
-};
-
-// CONTENT CONTAINER
-const SuccesViewContainer = ({ swap, onClick, loading, children }) => {
-  const { gameEditionView } = useGameEditionContext();
-  return (
-    <Content gameEditionView={gameEditionView}>
-      <Label fontFamily="bold" geCenter geColor="yellow" labelStyle={{ marginTop: 16 }}>
-        Preview Successful!
-      </Label>
-      {!gameEditionView && <SuccessfullIcon />}
-
-      <TransactionsDetails>
-        {children}
-        <GasCost swap={swap} />
-      </TransactionsDetails>
-      <CustomButton
-        type="secondary"
-        buttonStyle={{
-          width: '100%',
-          marginTop: !gameEditionView && '16px',
-          marginBottom: gameEditionView && '16px',
-        }}
-        onClick={async () => {
-          await onClick();
-        }}
-        loading={loading}
-      >
-        Send Transaction
-      </CustomButton>
-    </Content>
-  );
-};
-
-const SuccessViewContainerGE = ({ leftItem, rightItem, infoItems, hideIcon, title }) => {
-  const { gameEditionView } = useGameEditionContext();
-  return (
-    <PreviewContainer gameEditionView={gameEditionView}>
-      {title && (
-        <GameEditionLabel center color="yellow" fontSize={24} style={{ marginBottom: 0 }}>
-          {title}
-        </GameEditionLabel>
-      )}
-      <Row className="sb" style={{ position: 'relative', paddingLeft: 16, paddingRight: 16 }}>
-        <PixeledInfoContainerWhite>{leftItem}</PixeledInfoContainerWhite>
-        {!hideIcon && (
-          <PixeledCircleArrowIcon
-            style={{ position: 'absolute', width: 51.5, height: 49, top: 'calc(50% - 20px)', left: '44%', transform: 'rotate(-90deg)' }}
-          />
-        )}
-        <PixeledInfoContainerWhite>{rightItem}</PixeledInfoContainerWhite>
-      </Row>
-      <InfoContainer style={{ width: GE_DESKTOP_CONFIGURATION.displayWidth }}>
-        {infoItems?.map((item, i) => (
-          <PixeledBlueContainer key={i} label={item.label} value={item.value} />
-        ))}
-      </InfoContainer>
-      <Row className="c">
-        <PressButtonToActionLabel actionLabel="send" />
-      </Row>
-    </PreviewContainer>
   );
 };
